Remove: contact.py -> user.py to manage contact
