+++ app.py
@@ -18,9 +18,9 @@ from linebot.models import (
 from werkzeug.middleware.proxy_fix import ProxyFix
 
 import models.catcher as catcher
-import models.contact as contact
 import models.status as st
 import models.session as ss
+import models.user as user
 import config
 from database.database import init_db, db, get_db_uri
 import const.message as ms
@@ -65,13 +65,6 @@ schedule.every(1).week.do(cs.refresh)
 def test():
     users = models.User.query.all()
     print(users)
-    return Response(json.dumps({"status": "OK"}), mimetype='application/json')
-
-@app.route('/test/contact', methods=["GET"])
-def test_contact():
-    con = contact.Contact("0000", "0001")
-    db.session.add(con)
-    db.session.commit()
     return Response(json.dumps({"status": "OK"}), mimetype='application/json')
 
 
@@ -148,7 +141,7 @@ def handle_text_message(event):
         handle_catcher_rec(event)
     elif ss_type == st.Type.CONTACT:
         profile = line_bot_api.get_profile(user_id)
-        slack.send_msg_to_thread(profile.display_name, text, contact.get_thread(user_id))
+        slack.send_msg_to_thread(profile.display_name, text, user.get_thread(user_id))
     elif text == '次':
         handle_next(event)
     elif text in (ms.MSG_BN_CREATE_3_1, ms.MSG_BN_CREATE_3_2, ms.MSG_BN_CREATE_3_3, ms.MSG_BN_CREATE_3_5):
@@ -191,8 +184,7 @@ def handle_ctx0(event):
         s.set_context(user_id, 1)
         s.set_type(user_id, st.Type.CONTACT)
         res = slack.start_contact(profile.display_name)
-        print(type(res['message']['ts']))
-        contact.register(user_id, res['message']['ts'])
+        user.register(user_id, res['message']['ts'])
     else:
         line_bot_api.reply_message(event.reply_token, ms.MSG_DEFAULT)
 
@@ -293,7 +285,7 @@ def route_next(user_id: str):
 def reply_contact(event):
     if 'bot_id' not in event:
         thread_ts = event['thread_ts']
-        user_id = contact.get_user(thread_ts)
+        user_id = user.get_user(thread_ts)
         msg = event['text']
         line_bot_api.push_message(user_id, TextSendMessage(text=msg))
