+++ models/__init__.py
@@ -1,8 +1,6 @@
 from .user import User
-from .contact import Contact
 
 
 __all__ = [
 	User,
-	Contact,
 ]
